Refactor: Split Credits page into components
