+++ src/pages/collection/series/SeriesCredits.tsx
@@ -1,12 +1,11 @@
-import React, { useEffect, useMemo, useState } from 'react';
-import { useOutletContext, useParams } from 'react-router';
+import React, { useMemo, useState } from 'react';
+import { useParams } from 'react-router';
 import { mdiInformationOutline, mdiMagnify, mdiPlayCircleOutline } from '@mdi/js';
 import Icon from '@mdi/react';
-import { useVirtualizer } from '@tanstack/react-virtual';
 import { map, toNumber } from 'lodash';
 import { useDebounceValue } from 'usehooks-ts';
 
-import CharacterImage from '@/components/CharacterImage';
+import StaffPanelVirtualizer from '@/components/Collection/Credits/CreditsStaffVirtualizer';
 import Checkbox from '@/components/Input/Checkbox';
 import Input from '@/components/Input/Input';
 import MultiStateButton from '@/components/Input/MultiStateButton';
@@ -21,83 +20,13 @@ import useEventCallback from '@/hooks/useEventCallback';
 
 import type { SeriesCast } from '@/core/types/api/series';
 
-type ModeType = 'Character' | 'Staff';
-
-const getThumbnailUrl = (item: SeriesCast, mode: ModeType) => {
-  const thumbnail = item[mode].Image ?? null;
-  if (thumbnail === null) return null;
-  return `/api/v3/Image/${thumbnail.Source}/${thumbnail.Type}/${thumbnail.ID}`;
-};
-
-const creditTypeVariations = {
-  Character: 'Characters',
-  Staff: 'Staff',
-};
-
-const CreditsStaffPanel = React.memo(({ cast, mode }: { cast: SeriesCast, mode: ModeType }) => (
-  <div className="flex w-full flex-row items-center gap-6 rounded-lg border border-panel-border bg-panel-background-transparent p-6 font-semibold">
-    <div className="z-10 flex gap-x-2">
-      {mode === 'Character' && (
-        <CharacterImage
-          imageSrc={getThumbnailUrl(cast, 'Character')}
-          className="relative h-[7.75rem] w-[6.063rem] rounded-lg"
-        />
-      )}
-      <CharacterImage
-        imageSrc={getThumbnailUrl(cast, 'Staff')}
-        className="relative h-[7.75rem] w-[6.063rem] rounded-lg"
-      />
-    </div>
-    <div className="grow text-center">
-      <div className="line-clamp-2 text-base leading-8 xl:text-xl" title={cast[mode]?.Name}>
-        {cast[mode]?.Name}
-      </div>
-      {mode === 'Character' && <div className="opacity-65">{cast.Staff?.Name}</div>}
-      <div className="mt-2 text-sm">{cast.RoleDetails}</div>
-    </div>
-  </div>
-));
-
-const isCharacter = (item: SeriesCast) => item.RoleName === 'Seiyuu';
+export type CreditsModeType = 'Character' | 'Staff';
 
 const cleanString = (input = '') => input.replaceAll(' ', '').toLowerCase();
 
-const getUniqueDescriptions = (castList: SeriesCast[]) => [...new Set(castList.map(c => c.RoleDetails))];
+const getUniqueRoles = (castList: SeriesCast[]) => [...new Set(castList.map(c => c.RoleDetails))];
 
-const StaffPanelVirtualizer = ({ castArray, mode }: { castArray: SeriesCast[], mode: ModeType }) => {
-  const { scrollRef } = useOutletContext<{ scrollRef: React.RefObject<HTMLDivElement> }>();
-  const cardSize = { x: 466.5, y: 174, gap: 24 };
-
-  const rowVirtualizer = useVirtualizer({
-    count: castArray.length,
-    getScrollElement: () => scrollRef.current,
-    estimateSize: () => cardSize.y,
-    overscan: 30, // Greater than the norm as lanes aren't taken into account
-    lanes: 3,
-    gap: cardSize.gap,
-  });
-
-  return (
-    <div className="relative w-full" style={{ height: rowVirtualizer.getTotalSize() }}>
-      {rowVirtualizer.getVirtualItems().map(virtualRow => (
-        <div
-          key={virtualRow.key}
-          className="absolute top-0"
-          style={{
-            left: virtualRow.lane * (cardSize.x + cardSize.gap),
-            width: cardSize.x,
-            height: cardSize.y,
-            transform: `translateY(${virtualRow.start}px)`,
-          }}
-        >
-          <CreditsStaffPanel cast={castArray[virtualRow.index]} mode={mode} />
-        </div>
-      ))}
-    </div>
-  );
-};
-
-const states: { label?: string, value: ModeType }[] = [
+const modeStates: { label?: string, value: CreditsModeType }[] = [
   { label: 'Characters', value: 'Character' },
   { value: 'Staff' },
 ];
@@ -108,52 +37,50 @@ const SeriesCredits = () => {
   const { isPending: pendingRefreshAniDb, mutate: refreshAniDb } = useRefreshSeriesAniDBInfoMutation();
   const { isPending: pendingRefreshTvDb, mutate: refreshTvDb } = useRefreshSeriesTvdbInfoMutatation();
 
-  const [mode, setMode] = useState<ModeType>(states[0].value);
-  const handleModeChange = useEventCallback((newMode: ModeType) => {
-    setMode(newMode);
-  });
+  const [mode, setMode] = useState<CreditsModeType>(modeStates[0].value);
 
   const [search, setSearch] = useState('');
-
   const [debouncedSearch] = useDebounceValue(() => cleanString(search), 200);
 
-  const cast = useSeriesCastQuery(toNumber(seriesId!), !!seriesId).data;
-  const castByType = useMemo(() => ({
-    Character: cast?.filter(credit => isCharacter(credit)) ?? [],
-    Staff: cast?.filter(credit => !isCharacter(credit)) ?? [],
-  }), [cast]);
+  const [roleFilter, setRoleFilter] = useState<Set<string>>(new Set());
 
-  const uniqueDescriptions = useMemo(() => ({
-    Character: getUniqueDescriptions(castByType.Character),
-    Staff: getUniqueDescriptions(castByType.Staff),
-  }), [castByType]);
-
-  const [descriptionFilter, setDescriptionFilter] = useState<Set<string>>(new Set());
-
-  const filteredCast = useMemo(() => (castByType[mode].filter(p => (
-    (debouncedSearch === ''
-      || !!(cleanString(p?.Character?.Name).match(debouncedSearch))
-      || !!(cleanString(p?.Staff?.Name).match(debouncedSearch)))
-    && !descriptionFilter.has(p?.RoleDetails)
-  )).sort((a, b) => {
-    if (a[mode].Name > b[mode].Name) return 1;
-    if (a[mode].Name < b[mode].Name) return -1;
-    return 0;
-  })), [castByType, mode, debouncedSearch, descriptionFilter]);
-
-  useEffect(() => {
-    setSearch('');
-    setDescriptionFilter(new Set());
-  }, [mode]);
+  const handleModeChange = useEventCallback((newMode: CreditsModeType) => {
+    setMode(() => {
+      setSearch('');
+      setRoleFilter(new Set());
+      return newMode;
+    });
+  });
 
   const handleFilterChange = useEventCallback((event: React.ChangeEvent<HTMLInputElement>) => {
     const { id: description } = event.target;
-    setDescriptionFilter((prevState) => {
+    setRoleFilter((prevState) => {
       const newState = new Set(prevState);
       if (!newState.delete(description)) newState.add(description);
       return newState;
     });
   });
+
+  const cast = useSeriesCastQuery(toNumber(seriesId!), !!seriesId).data;
+  const castByType = useMemo(() => ({
+    Character: cast?.filter(credit => credit.RoleName === 'Seiyuu') ?? [],
+    Staff: cast?.filter(credit => credit.RoleName !== 'Seiyuu') ?? [],
+  }), [cast]);
+
+  const uniqueRoles = useMemo(() => ({
+    Character: getUniqueRoles(castByType.Character),
+    Staff: getUniqueRoles(castByType.Staff),
+  }), [castByType]);
+
+  const filteredCast = useMemo(() => (castByType[mode].filter(p => (
+    (debouncedSearch === ''
+      || !!([p?.Character?.Name, p?.Staff?.Name].some(name => cleanString(name).match(debouncedSearch))))
+    && !roleFilter.has(p?.RoleDetails)
+  )).sort((a, b) => {
+    if (a[mode].Name > b[mode].Name) return 1;
+    if (a[mode].Name < b[mode].Name) return -1;
+    return 0;
+  })), [castByType, mode, debouncedSearch, roleFilter]);
 
   if (!seriesId) return null;
 
@@ -185,13 +112,13 @@ const SeriesCredits = () => {
           <div className="flex flex-col gap-y-2">
             <div className="text-base font-semibold">Roles</div>
             <div className="flex flex-col gap-y-2 rounded-lg bg-panel-input p-6">
-              {map(uniqueDescriptions[mode], desc => (
+              {map(uniqueRoles[mode], desc => (
                 <Checkbox
                   justify
                   label={desc}
                   key={desc}
                   id={desc}
-                  isChecked={!descriptionFilter.has(desc)}
+                  isChecked={!roleFilter.has(desc)}
                   onChange={handleFilterChange}
                 />
               ))}
@@ -246,7 +173,7 @@ const SeriesCredits = () => {
         <div className="flex h-[6.125rem] items-center justify-between rounded-lg border border-panel-border bg-panel-background-transparent px-6 py-4">
           <div className="text-xl font-semibold">
             Credits |&nbsp;
-            {(debouncedSearch !== '' || descriptionFilter.size > 0) && (
+            {(debouncedSearch !== '' || roleFilter.size > 0) && (
               <>
                 <span className="text-panel-text-important">
                   {filteredCast.length}
@@ -258,10 +185,10 @@ const SeriesCredits = () => {
               {castByType[mode].length ?? 0}
             </span>
             &nbsp;
-            {creditTypeVariations[mode]}
+            {mode === 'Character' ? 'Characters' : mode}
             &nbsp;Listed
           </div>
-          <MultiStateButton activeState={mode} states={states} onStateChange={handleModeChange} />
+          <MultiStateButton activeState={mode} states={modeStates} onStateChange={handleModeChange} />
         </div>
         <StaffPanelVirtualizer castArray={filteredCast} mode={mode} />
       </div>
